Turn keyvalue_dictionary plugin into a utility.
