+++ lib/layer/_layer.mjs
@@ -34,7 +34,6 @@ import basic from './themes/basic.mjs';
 import categorized from './themes/categorized.mjs';
 import distributed from './themes/distributed.mjs';
 import graduated from './themes/graduated.mjs';
-import { parseObject } from '../plugins/keyvalue_dictionary.mjs';
 
 export default {
   decorate,
@@ -44,7 +43,6 @@ export default {
   featureHover,
   featureStyle,
   fade,
-  keyvalue_dictionary,
   styleParser,
   Style,
   themes: {
@@ -71,19 +69,4 @@ function Style(layer) {
     `The mapp.layer.Style() method has been superseeded by the mapp.layer.featureStyle() method.`,
   );
   return featureStyle(layer);
-}
-
-/**
-@function keyvalue_dictionary
-
-@description
-Replaces key-value pairs in the layer object with dictionary entries. Will be executed instead of the keyvalue_dictionary plugin method from the layer decorator.
-
-@param {layer} layer A decorated layer object.
-@property {Object} layer.keyvalue_dictionary The keyvalue dictionary.
-*/
-export function keyvalue_dictionary(layer) {
-  if (!layer?.keyvalue_dictionary) return;
-
-  parseObject(layer, layer.keyvalue_dictionary);
 }
