+++ lib/layer/decorate.mjs
@@ -16,6 +16,8 @@ Any plugins matching layer keys will be executed with the layer being passed as 
 
 The layer object is returned from the decorator.
 
+@requires /utils/keyvalue_dictionary
+
 @module /layer/decorate
 */
 
@@ -32,6 +34,9 @@ The layer decorator method create mapp-layer typedef object from a json-layer.
 @returns {layer} Decorated Mapp Layer.
 */
 export default async function decorate(layer) {
+  // Replace keyvalue_dictionary entries if required.
+  mapp.utils.keyvalue_dictionary(layer);
+
   // Check layer format.
   if (!Object.hasOwn(mapp.layer.formats, layer.format)) {
     console.warn(`Layer: ${layer.key}; ${layer.format} format unavailable.`);
@@ -313,4 +318,19 @@ function changeEnd(layer) {
   if (layer.params.viewport) {
     layer.display && layer.reload();
   }
+}
+
+/**
+@function keyvalue_dictionary
+
+@description
+Replaces key-value pairs in the layer object with dictionary entries. Will be executed instead of the keyvalue_dictionary plugin method from the layer decorator.
+
+@param {layer} layer A decorated layer object.
+@property {Object} layer.keyvalue_dictionary The keyvalue dictionary.
+*/
+export function keyvalue_dictionary(layer) {
+  if (!layer?.keyvalue_dictionary) return;
+
+  parseObject(layer, layer.keyvalue_dictionary);
 }
