+++ lib/mapview/_mapview.mjs
@@ -63,6 +63,14 @@ The mapview decorator may return the async mapviewPromise method which must be a
 @returns {mapview} Decorated Mapview.
 */
 export default function decorate(mapview) {
+  if (!mapview.locale) {
+    console.warn(`A locale property is required to decorare a mapview object.`);
+    return;
+  }
+
+  // Replace keyvalue_dictionary entries if required.
+  mapp.utils.keyvalue_dictionary(mapview.locale);
+
   // Assign host from mapp if not implicit.
   mapview.host ??= mapp.host;
