+++ lib/plugins/_plugins.mjs
@@ -8,7 +8,6 @@ The module exports a collection of core mapp plugins.
 @requires module:/plugins/admin
 @requires module:/plugins/feature_info
 @requires module:/plugins/fullscreen
-@requires module:/plugins/keyvalue_dictionary
 @requires module:/plugins/link_button
 @requires module:/plugins/locator
 @requires module:/plugins/login
@@ -23,7 +22,6 @@ import { admin } from './admin.mjs';
 import { dark_mode } from './dark_mode.mjs';
 import { feature_info } from './feature_info.mjs';
 import { fullscreen } from './fullscreen.mjs';
-import { keyvalue_dictionary } from './keyvalue_dictionary.mjs';
 import { link_button } from './link_button.mjs';
 import { locator } from './locator.mjs';
 import { login } from './login.mjs';
@@ -38,7 +36,6 @@ const plugins = {
   dark_mode,
   feature_info,
   fullscreen,
-  keyvalue_dictionary,
   link_button,
   locator,
   login,
