+++ lib/utils/_utils.mjs
@@ -33,6 +33,8 @@ import { default as hexa } from './hexa.mjs';
 
 import jsonParser from './jsonParser.mjs';
 
+import { keyvalue_dictionary } from './keyvalue_dictionary.mjs';
+
 import loadPlugins from './loadPlugins.mjs';
 
 import getCurrentPosition from './getCurrentPosition.mjs';
@@ -96,6 +98,7 @@ export default {
   getCurrentPosition,
   hexa,
   jsonParser,
+  keyvalue_dictionary,
   loadPlugins,
   merge,
   mobile,
