+++ lib/utils/keyvalue_dictionary.mjs
@@ -1,7 +1,9 @@
 /**
-### /plugins/keyvalue_dictionary
+### /utils/keyvalue_dictionary
 
-A keyvalue_dictionary can be configured in the locale. The keyvalue_dictionary plugin method will parse the mapview.locale for properties with key/value pairs matching entries in the dictionary. The property value will be replaced with a language lookup in the dictionary or the default value.
+The keyvalue_dictionary utility method will parse a JSON object for properties with key/value pairs matching entries a provided dictionary.
+
+The property value will be replaced with a language lookup in the dictionary or the default value.
 
 ```js
 "keyvalue_dictionary": [
@@ -13,7 +15,7 @@ A keyvalue_dictionary can be configured in the locale. The keyvalue_dictionary p
   }
 ]
 ```
-@module /plugins/keyvalue_dictionary
+@module /utils/keyvalue_dictionary
 */
 
 /**
@@ -26,16 +28,16 @@ Replaces key-value pairs in the mapview locale object with dictionary entries.
 @param {mapview} mapview The mapview object.
 @property {Object} mapview.locale The locale object of the mapview.
 */
-export function keyvalue_dictionary(keyvalue_dictionary, mapview) {
-  if (!mapview?.locale) return;
+export function keyvalue_dictionary(obj) {
+  if (!obj.keyvalue_dictionary) return;
 
   const dictionary = new WeakMap();
 
-  keyvalue_dictionary.forEach((entry) => {
+  obj.keyvalue_dictionary.forEach((entry) => {
     dictionary.set([entry.key, entry.value], entry);
   });
 
-  parseObject(mapview.locale, dictionary);
+  parseObject(obj, dictionary);
 }
 
 /**
@@ -47,11 +49,9 @@ Parses an object and replaces its string values with dictionary entries.
 @param {Object} obj The object to parse.
 @param {Array} dictionary An array of dictionary entries.
 */
-export function parseObject(obj, dictionary) {
+function parseObject(obj, dictionary) {
   for (const [key, value] of Object.entries(obj)) {
-    // Prevents crash where the mapview with the locale itself maybe nested in a locale plugin object.
     if (key === 'mapview') continue;
-
     if (isArrayObject(value, dictionary)) continue;
     if (typeof value === 'string') replaceKeyValue(obj, key, value, dictionary);
   }
